% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OandaForexLabs.R
\name{Autochartist.Oanda}
\alias{Autochartist.Oanda}
\title{Returns ‘Our Favourites’ signals from Autochartist.}
\usage{
Autochartist.Oanda(ACCOUNT_TYPE, ACCESS_TOKEN, INSTRUMENTS = "EUR_USD",
  PERIOD = "1 week")
}
\arguments{
\item{ACCOUNT_TYPE}{"practice", "real" or "sandbox"}

\item{ACCESS_TOKEN}{The authorization bearer token previously obtained by the client, can be found in oanda environment}

\item{INSTRUMENTS}{Instrument to get candlestick data for e.g. "AUD_USD"}

\item{PERIOD}{Required Period of time to retrieve calendar data for.}
}
\value{
A \code{list}
}
\description{
Returns ‘Our Favourites’ signals from Autochartist. These signals can be a Chart Pattern, or a Key Level.
Chart Patterns consists of an upper and a lower line forming a pattern, the pattern’s end time, and the prediction price box.
Key Levels have only a price level.
}
\details{
Period includes:\cr
 - 4 hours\cr
 - 8 hours\cr
 - 12 hours\cr
 - 1 day\cr
 - 1 week\cr
}
\examples{
signals = Autochartist.Oanda(.oandaEnv$ACCOUNT_TYPE, .oandaEnv$ACCESS_TOKEN, INSTRUMENTS = 'AUD_USD', PERIOD = '1 week')
signals

}
