% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBrokersInstrument.R
\name{cancelIBFundamentalData}
\alias{cancelIBFundamentalData}
\title{A Wrapper for IB Fundamental Data Cancellation}
\usage{
cancelIBFundamentalData(twsconn, reqId)
}
\arguments{
\item{twsconn}{connection to current TWS account, if NULL, a new connection will be created within the function}

\item{reqId}{tws request ID}
}
\description{
A Wrapper for IB Fundamental Data Cancellation
}
