% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBrokersInstrument.R
\name{cancelIBNewsBulletins}
\alias{cancelIBNewsBulletins}
\title{Cancel News Bulletins of Interactive Brokers}
\usage{
cancelIBNewsBulletins(twsconn)
}
\arguments{
\item{twsconn}{connection to current TWS account, if NULL, a new connection will be created within the function}
}
\description{
Cancel News Bulletins of Interactive Brokers
}
