% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OandaOrders.R
\name{cancelOandaOrder}
\alias{cancelOandaOrder}
\title{Cancel a pending Oanda offer}
\usage{
cancelOandaOrder(ACCOUNT_TYPE, ACCESS_TOKEN, ACCOUNT_ID, ORDER)
}
\arguments{
\item{ACCOUNT_TYPE}{"practice", "real" or "sandbox"}

\item{ACCESS_TOKEN}{The authorization bearer token previously obtained by the client, can be found in oanda environment}

\item{ACCOUNT_ID}{ID of the Account to create the Order for.}

\item{ORDER}{order id}
}
\value{
A \code{HTTP Response}
}
\description{
Cancel a pending Oanda offer
}
