% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleIBrokersInterruption.R
\name{checkIBConnection}
\alias{checkIBConnection}
\title{Notification of IB Gateway Interruption}
\usage{
checkIBConnection(toAddress = c(
  "ivan.liuyanfeng@gmail.com"))
}
\description{
Notification of IB Gateway Interruption
}
\examples{
checkIBConnection()

}
