% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OandaTrades.R
\name{cloaseOandaTrade}
\alias{cloaseOandaTrade}
\title{Close a trade in Oanda}
\usage{
cloaseOandaTrade(ACCOUNT_TYPE, ACCESS_TOKEN, ACCOUNT_ID, TRADE_ID, ...)
}
\arguments{
\item{ACCOUNT_TYPE}{"practice", "real" or "sandbox"}

\item{ACCESS_TOKEN}{The authorization bearer token previously obtained by the client, can be found in oanda environment}

\item{ACCOUNT_ID}{ID of the Account to create the Order for.}

\item{TRADE_ID}{Identifier of trade to be closed}
}
\description{
Close a trade in Oanda
}
\examples{
cloaseOandaTrade(.oandaEnv$ACCOUNT_TYPE, .oandaEnv$ACCESS_TOKEN, .oandaEnv$ACCOUNT_ID, "XXXXX")

}
