% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OandaForexLabs.R
\name{COT.Oanda}
\alias{COT.Oanda}
\title{Commitments of Traders data from the CFTC for supported currencies}
\usage{
COT.Oanda(ACCOUNT_TYPE, ACCESS_TOKEN, INSTRUMENTS = "EUR_USD")
}
\arguments{
\item{ACCOUNT_TYPE}{"practice", "real" or "sandbox"}

\item{ACCESS_TOKEN}{The authorization bearer token previously obtained by the client, can be found in oanda environment}

\item{INSTRUMENTS}{Instrument to get candlestick data for e.g. "AUD_USD"}
}
\value{
A \code{data.table}
}
\description{
Returns up to 4 years worth of Commitments of Traders data from the CFTC for supported currencies.
This is essentially CFTC’s Non-Commercial order book data.
}
\examples{
cot = COT.Oanda(.oandaEnv$ACCOUNT_TYPE, .oandaEnv$ACCESS_TOKEN, INSTRUMENTS = 'AUD_USD')
cot

}
