% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OandaOrders.R
\name{createOandaPositionClose}
\alias{createOandaPositionClose}
\title{Close out a Oanda Position}
\usage{
createOandaPositionClose(ACCOUNT_TYPE, ACCESS_TOKEN, ACCOUNT_ID,
  INSTRUMENTS = "AUD_USD", LONGUNITS = NULL, SHORTUNITS = NULL)
}
\arguments{
\item{ACCOUNT_TYPE}{"practice", "real" or "sandbox"}

\item{ACCESS_TOKEN}{The authorization bearer token previously obtained by the client, can be found in oanda environment}

\item{ACCOUNT_ID}{ID of the Account to create the Order for.}

\item{INSTRUMENTS}{Instrument to get candlestick data for e.g. "AUD_USD"}

\item{LONGUNITS}{Number of units of the instrument to be executed, 'ALL' or numbers}

\item{SHORTUNITS}{Number of units of the instrument to be executed}
}
\value{
A \code{HTTP Response}
}
\description{
Close out a Oanda Position
}
\examples{
createOandaPositionClose(.oandaEnv$ACCOUNT_TYPE, .oandaEnv$ACCESS_TOKEN, .oandaEnv$ACCOUNT_ID, INSTRUMENTS = 'AUD_USD', LONGUNITS = 'ALL')
# HTTP 201 – The Order was created as specified
# Response [https://api-fxpractice.oanda.com/v3/accounts/101-011-4686012-003/orders]
# Date: 2016-11-28 11:38
# Status: 201
# Content-Type: application/json
# Size: 775 B

}
