% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OandaOrders.R
\name{createOandaTPSLOrder}
\alias{createOandaTPSLOrder}
\title{Create Stoploss or Takeprofit Order for an Oanda Account}
\usage{
createOandaTPSLOrder(ACCOUNT_TYPE, ACCESS_TOKEN, ACCOUNT_ID, TRADEID,
  PRICE = NULL, ORDERTYPE = "TAKE_PROFIT", ...)
}
\arguments{
\item{ACCOUNT_TYPE}{"practice", "real" or "sandbox"}

\item{ACCESS_TOKEN}{The authorization bearer token previously obtained by the client, can be found in oanda environment}

\item{ACCOUNT_ID}{ID of the Account to create the Order for.}

\item{TRADEID}{trade id}

\item{PRICE}{prices to trigger the action}

\item{ORDERTYPE}{"TAKE_PROFIT", "STOP_LOSS"}
}
\value{
A \code{HTTP Response}
}
\description{
Create Stoploss or Takeprofit Order for an Oanda Account
}
\examples{
createOandaTPSLOrder(.oandaEnv$ACCOUNT_TYPE, .oandaEnv$ACCESS_TOKEN, .oandaEnv$ACCOUNT_ID, TRADEID = '1111', PRICE = 1.3475, ORDERTYPE = 'TAKE_PROFIT')
# HTTP 201 – The Order was created as specified
# Response [https://api-fxpractice.oanda.com/v3/accounts/101-011-4686012-003/orders]
# Date: 2016-11-28 11:38
# Status: 201
# Content-Type: application/json
# Size: 775 B

}
