% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconomicCalendar.R
\name{getHeadlinesDailyFX}
\alias{getHeadlinesDailyFX}
\title{Fetch Forex News Feed from DailyFX}
\usage{
getHeadlinesDailyFX(NEWS_TYPE = "trading_news_events")
}
\arguments{
\item{NEWS_TYPE}{the types of the latest DailyFX headlines on the topics that interest you. Default "trading_news_events"}
}
\value{
A \code{data.table} of the latest DailyFX headlines
}
\description{
Fetch Forex News Feed from DailyFX
}
\details{
News Types including (Default "trading_news_events" for the function): \cr
 - All\cr
 - ForexNewsEvents\cr
 - TradingNewsEvents\cr
 - Forecasts\cr
 - SpecialReports\cr
 - TechnicalArticles\cr
 - Sentiment\cr
 - SpecialReports\cr
 - Alerts\cr
 - AnalystPicks\cr
 - DailyTechnicals\cr
 - Gold\cr
 - Oil\cr
}
\examples{
dat = getHeadlinesDailyFX(NEWS_TYPE = "Sentiment")
View(dat)

}
