% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OandaPricing.R
\name{getOandaCurPricing}
\alias{getOandaCurPricing}
\title{Get Priceing Information for Oanda Instruments}
\usage{
getOandaCurPricing(ACCOUNT_TYPE, ACCESS_TOKEN, ACCOUNT_ID,
  INSTRUMENTS = c("AUD_USD"))
}
\arguments{
\item{ACCOUNT_TYPE}{"practice", "real" or "sandbox"}

\item{ACCESS_TOKEN}{The authorization bearer token previously obtained by the client, can be found in oanda environment}

\item{ACCOUNT_ID}{ID of the Account to create the Order for.}

\item{INSTRUMENTS}{Instrument to get candlestick data for e.g. "AUD_USD"}
}
\value{
A \code{data.frame} of current prices
}
\description{
Get pricing information for a specified list of Instruments within an Account
}
\examples{
getOandaCurPricing(.oandaEnv$ACCOUNT_TYPE, .oandaEnv$ACCESS_TOKEN, .oandaEnv$ACCOUNT_ID, INSTRUMENTS = c('AUD_USD'))

}
