% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OandaInstrument.R
\name{getOandaInstrumentCandles}
\alias{getOandaInstrumentCandles}
\title{Fetch Candlestick Data for an Instrument from Oanda}
\usage{
getOandaInstrumentCandles(ACCOUNT_TYPE, ACCESS_TOKEN, INSTRUMENTS = "AUD_USD",
  count = 500, price = "BA", granularity = "S5", ...)
}
\arguments{
\item{ACCOUNT_TYPE}{"practice", "real" or "sandbox"}

\item{ACCESS_TOKEN}{The authorization bearer token previously obtained by the client, can be found in oanda environment}

\item{INSTRUMENTS}{Instrument to get candlestick data for e.g. "AUD_USD"}

\item{count}{The number of candlesticks to return in the reponse. Count should not be specified if both the start and end parameters are provided}

\item{price}{The Price component(s) to get candlestick data for. Can contain any combination of the characters “M” (midpoint candles) “B” (bid candles) and “A” (ask candles). [default=M]}

\item{granularity}{The granularity of the candlesticks to fetch [default=S5]}
}
\value{
A \code{data.frame} of candlesticks
}
\description{
Fetch Candlestick Data for an Instrument from Oanda
}
\examples{
dat = getOandaInstrumentCandles(.oandaEnv$ACCOUNT_TYPE, .oandaEnv$ACCESS_TOKEN, INSTRUMENTS = 'AUD_USD', price = 'BA', granularity = 'M5', count = 500)

}
