% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OandaOrders.R
\name{getOandaOrderList}
\alias{getOandaOrderList}
\title{Get a list of Orders of an Account}
\usage{
getOandaOrderList(ACCOUNT_TYPE, ACCESS_TOKEN, ACCOUNT_ID,
  INSTRUMENTS = "AUD_USD", PENDING = FALSE)
}
\arguments{
\item{ACCOUNT_TYPE}{"practice", "real" or "sandbox"}

\item{ACCESS_TOKEN}{The authorization bearer token previously obtained by the client, can be found in oanda environment}

\item{ACCOUNT_ID}{ID of the Account to create the Order for.}

\item{INSTRUMENTS}{Instrument to get candlestick data for e.g. "AUD_USD"}

\item{PENDING}{return only pending orders or not}
}
\value{
A \code{list} of order histories
}
\description{
Get a list of Orders of an Account
}
\examples{
getOandaOrderList(.oandaEnv$ACCOUNT_TYPE, .oandaEnv$ACCESS_TOKEN, .oandaEnv$ACCOUNT_ID, INSTRUMENTS = 'USD_CAD')
# $lastTransactionID
# [1] "2584"
# $orders
# list()

}
