% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OandaPositions.R
\name{getOandaPositions}
\alias{getOandaPositions}
\title{List all Positions for an Account}
\usage{
getOandaPositions(ACCOUNT_TYPE, ACCESS_TOKEN, ACCOUNT_ID)
}
\arguments{
\item{ACCOUNT_TYPE}{"practice", "real" or "sandbox"}

\item{ACCESS_TOKEN}{The authorization bearer token previously obtained by the client, can be found in oanda environment}

\item{ACCOUNT_ID}{ID of the Account to create the Order for.}
}
\value{
A \code{list} of account positions
}
\description{
List all Positions for an Account. The Positions returned are for every instrument that has
had a position during the lifetime of an the Account.
}
\examples{
getOandaPositions(.oandaEnv$ACCOUNT_TYPE, .oandaEnv$ACCESS_TOKEN, .oandaEnv$ACCOUNT_ID)

}
