% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OandaTrades.R
\name{getOandaTrades}
\alias{getOandaTrades}
\title{Get a list of Trades for an Oanda Account}
\usage{
getOandaTrades(ACCOUNT_TYPE, ACCESS_TOKEN, ACCOUNT_ID,
  INSTRUMENTS = "AUD_USD", OPEN = FALSE)
}
\arguments{
\item{ACCOUNT_TYPE}{"practice", "real" or "sandbox"}

\item{ACCESS_TOKEN}{The authorization bearer token previously obtained by the client, can be found in oanda environment}

\item{ACCOUNT_ID}{ID of the Account to create the Order for.}

\item{INSTRUMENTS}{Instrument to get candlestick data for e.g. "AUD_USD"}

\item{OPEN}{Show only open trades or not}
}
\value{
A \code{data.frame} of trades
}
\description{
Get a list of Trades for an Oanda Account
}
\examples{
getOandaTrades(.oandaEnv$ACCOUNT_TYPE, .oandaEnv$ACCESS_TOKEN, .oandaEnv$ACCOUNT_ID, INSTRUMENTS = c('AUD_USD'))

}
