% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OandaTransactions.R
\name{getOandaTxns}
\alias{getOandaTxns}
\title{Get a list of Transactions of Oanda Account}
\usage{
getOandaTxns(ACCOUNT_TYPE, ACCESS_TOKEN, ACCOUNT_ID, PAGESIZE = 1000,
  FROM = NULL, TO = NULL)
}
\arguments{
\item{ACCOUNT_TYPE}{"practice", "real" or "sandbox"}

\item{ACCESS_TOKEN}{The authorization bearer token previously obtained by the client, can be found in oanda environment}

\item{ACCOUNT_ID}{ID of the Account to create the Order for.}
}
\value{
A \code{data.frame} of transactions
}
\description{
Get the details of a single Oanda Account Transactions
}
\examples{
getOandaTxns(.oandaEnv$ACCOUNT_TYPE, .oandaEnv$ACCESS_TOKEN, .oandaEnv$ACCOUNT_ID)

}
