% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAPI.R
\name{IG_close_open_pos}
\alias{IG_close_open_pos}
\title{IG API Closes one or more OTC positions}
\usage{
IG_close_open_pos(headers,
  url = "https://demo-api.ig.com/gateway/deal/positions/otc", deal_id = "",
  direction, epic, expiry = "-", level = "", timeInForce = "",
  order_type = "MARKET", quote_id = "", size, timeo = 5)
}
\arguments{
\item{headers}{Object returned from \code{IG_Auth}}

\item{url}{API URL}

\item{deal_id}{Deal identifier}

\item{direction}{Deal direction ('BUY' or 'SELL')}

\item{epic}{Instrument epic identifier}

\item{expiry}{Instrument expiry}

\item{level}{Closing deal level}

\item{order_type}{'LIMIT', 'MARKET', 'QUATE'}

\item{quote_id}{Lightstreamer price quote identifier}

\item{size}{Deal size}

\item{timeo}{number of tries}
}
\value{
A \code{data.frame} Close position response
}
\description{
Closes one or more OTC positions
}
\examples{
HEADERS = IG_Auth(" ","APIdemo1", " ")
IG_close_open_pos(headers = HEADERS, direction = 'BUY', epic = 'CS.D.AUDCAD.CFD.IP', size = 1)

}
