% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAPI.R
\name{IG_create_open_pos}
\alias{IG_create_open_pos}
\title{IG API Create one or more OTC positions}
\usage{
IG_create_open_pos(headers,
  url = "https://demo-api.ig.com/gateway/deal/positions/otc",
  dealReference = "", currency_code = "AUD", direction = "BUY", epic,
  expiry = "-", force_open = "true", guaranteed_stop = "false",
  level = "", limit_distance = "", limit_level = "",
  order_type = "MARKET", size, stop_distance = "", stop_level = "",
  trailingStop = "false", trailingStopIncrement = "",
  timeInForce = "FILL_OR_KILL", timeo = 5)
}
\arguments{
\item{headers}{Object returned from \code{IG_Auth}}

\item{url}{API URL}

\item{dealReference}{A user-defined reference identifying the submission of the order}

\item{currency_code}{Currency. Restricted to available instrument currencies}

\item{direction}{Deal direction ('BUY' or 'SELL')}

\item{epic}{Instrument epic identifier}

\item{expiry}{Instrument expiry}

\item{force_open}{True if force open is required}

\item{guaranteed_stop}{True if a guaranteed stop is required}

\item{level}{Closing deal level}

\item{limit_distance}{Limit distance}

\item{limit_level}{Limit level}

\item{order_type}{'LIMIT', 'MARKET', 'QUATE'}

\item{size}{Deal size}

\item{stop_distance}{Stop distance}

\item{stop_level}{Stop level}

\item{trailingStop}{Whether the stop has to be moved towards the current level in case of a favourable trade}

\item{trailingStopIncrement}{increment step in pips for the trailing stop}

\item{timeInForce}{'EXECUTE_AND_ELIMINATE' or 'FILL_OR_KILL'}

\item{timeo}{number of tries}
}
\value{
A \code{data.frame} 	Deal reference of the transaction
}
\description{
Create one or more OTC positions
}
\examples{
HEADERS = IG_Auth(" ","APIdemo1", " ")
order = IG_create_open_pos(headers = HEADERS, url ="https://demo-api.ig.com/gateway/deal/positions/otc",
dealReference = 'audcad001', currency_code = 'AUD', direction = 'BUY', epic = 'CS.D.AUDUSD.CFD.IP',
expiry = '-', force_open = 'true', guaranteed_stop = 'false', level = '', limit_distance = '', limit_level = '',
order_type = 'MARKET', size = 3,
stop_distance = 10, stop_level = '', trailingStop = 'false', trailingStopIncrement = '',
timeInForce = 'FILL_OR_KILL', timeo=5)

}
