% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAPI.R
\name{IG_create_working_order}
\alias{IG_create_working_order}
\title{IG API Creates an OTC working order.}
\usage{
IG_create_working_order(headers,
  url = "https://demo-api.ig.com/gateway/deal/workingorders/otc",
  dealReference, currency_code, direction, epic, expiry, good_till_date,
  guaranteed_stop, level, limit_distance, limit_level, size, stop_distance,
  stop_level, time_in_force, order_type, timeo = 5)
}
\arguments{
\item{headers}{Object returned from \code{IG_Auth}}

\item{url}{API URL}

\item{dealReference}{A user-defined reference identifying the submission of the order}

\item{currency_code}{Currency. Restricted to available instrument currencies}

\item{direction}{Deal direction ('BUY' or 'SELL')}

\item{epic}{Instrument epic identifier}

\item{expiry}{Instrument expiry}

\item{good_till_date}{This accepts two possible formats either yyyy/mm/dd hh:mm:ss in UTC Time or Unix Timestamp in milliseconds}

\item{guaranteed_stop}{True if a guaranteed stop is required}

\item{level}{Closing deal level}

\item{limit_distance}{Limit distance}

\item{limit_level}{Limit level}

\item{size}{Deal size}

\item{stop_distance}{Stop distance}

\item{stop_level}{Stop level}

\item{order_type}{'LIMIT', 'MARKET', 'QUATE'}

\item{timeo}{number of tries}

\item{timeInForce}{'EXECUTE_AND_ELIMINATE' or 'FILL_OR_KILL'}
}
\value{
A \code{data.frame} 	OTC working order.
}
\description{
Creates an OTC working order.
}
\examples{
HEADERS = IG_Auth(" ","APIdemo1", " ")
IG_create_working_order(HEADERS) # to be updated

}
