% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAPI.R
\name{IG_del_working_order}
\alias{IG_del_working_order}
\title{IG API Deletes an OTC working order.}
\usage{
IG_del_working_order(headers,
  url = "https://demo-api.ig.com/gateway/deal/workingorders/otc", deal_id,
  timeo = 5)
}
\arguments{
\item{headers}{Object returned from \code{IG_Auth}}

\item{url}{API URL}

\item{deal_id}{deal ID}

\item{timeo}{number of tries}
}
\value{
A \code{data.frame} Deletes OTC working order.
}
\description{
Deletes an OTC working order.
}
\examples{
HEADERS = IG_Auth(" ","APIdemo1", " ")
IG_del_working_order(HEADERS, deal_id = '') # to be updated

}
