% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAPI.R
\name{IG_fetch_account_act}
\alias{IG_fetch_account_act}
\title{IG API Fetch Account Activities}
\usage{
IG_fetch_account_act(headers,
  url = "https://demo-api.ig.com/gateway/deal/history/activity",
  milliseconds = 3600000, timeo = 5)
}
\arguments{
\item{headers}{Object returned from \code{IG_Auth}}

\item{url}{API URL}

\item{milliseconds}{lookback periods}

\item{timeo}{number of tries}
}
\value{
A \code{data.frame} of account activities
}
\description{
Fetch IG Account Activities
}
\examples{
HEADERS = IG_Auth(" ","APIdemo1", " ")
IG_fetch_account_act(HEADERS)

}
