% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAPI.R
\name{IG_fetch_accounts}
\alias{IG_fetch_accounts}
\title{IG API Fetch Account Information}
\usage{
IG_fetch_accounts(headers,
  url = "https://demo-api.ig.com/gateway/deal/accounts", timeo = 5)
}
\arguments{
\item{headers}{Object returned from \code{IG_Auth}}

\item{url}{API URL}

\item{timeo}{number of tries}
}
\value{
A \code{data.frame} of account details
}
\description{
Fetch IG Account Information
}
\examples{
HEADERS = IG_Auth(" ","APIdemo1", " ")
IG_fetch_accounts(HEADERS)

}
