% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAPI.R
\name{IG_fetch_all_watchlists}
\alias{IG_fetch_all_watchlists}
\title{IG API Get All Watchlists}
\usage{
IG_fetch_all_watchlists(headers,
  url = "https://demo-api.ig.com/gateway/deal/watchlists", timeo = 5)
}
\arguments{
\item{headers}{Object returned from \code{IG_Auth}}

\item{url}{API URL}

\item{timeo}{number of tries}
}
\value{
A \code{data.frame} of watchlists information
}
\description{
Get All Watchlists
}
\examples{
HEADERS = IG_Auth(" ","APIdemo1", " ")
IG_fetch_all_watchlists(HEADERS)

}
