% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAPI.R
\name{IG_fetch_deal}
\alias{IG_fetch_deal}
\title{IG API Returns a deal confirmation for the given deal reference}
\usage{
IG_fetch_deal(headers, url = "https://demo-api.ig.com/gateway/deal/confirms",
  deal_reference, timeo = 5)
}
\arguments{
\item{headers}{Object returned from \code{IG_Auth}}

\item{url}{API URL}

\item{deal_reference}{deal reference}
}
\value{
A \code{data.frame} deal confirmation
}
\description{
Returns a deal confirmation for the given deal reference
}
\examples{
HEADERS = IG_Auth(" ","APIdemo1", " ")
IG_fetch_deal(HEADERS, deal_reference = '')

}
