% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAPI.R
\name{IG_fetch_hist_prices}
\alias{IG_fetch_hist_prices}
\title{IG API Fetch Historical Prices}
\usage{
IG_fetch_hist_prices(epic, n = 50, res = "DAY", headers,
  url = "https://demo-api.ig.com/gateway/deal/prices", timeo = 5)
}
\arguments{
\item{epic}{Instrument ID}

\item{n}{Number of points to retrieve}

\item{res}{Granularity of data}

\item{headers}{Object returned from \code{IG_Auth}}

\item{url}{API URL}

\item{timeo}{number of tries}
}
\value{
A \code{data.frame} of historical prices
}
\description{
Fetch IG Account Historical Prices
}
\examples{
HEADERS = IG_Auth(" ","APIdemo1", " ")
audusd = IG_fetch_hist_prices("CS.D.AUDUSD.CFD.IP", 50, "HOUR", headers = HEADERS)
usdcad = IG_fetch_hist_prices("CS.D.USDCAD.CFD.IP", 50, "HOUR", headers = HEADERS)

}
