% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAPI.R
\name{IG_fetch_markets}
\alias{IG_fetch_markets}
\title{IG API Fetch Markets Information}
\usage{
IG_fetch_markets(epic = "IX.D.FTSE.IFM.IP", headers,
  url = "https://demo-api.ig.com/gateway/deal/markets", timeo = 5)
}
\arguments{
\item{epic}{Instrument ID}

\item{headers}{Object returned from \code{IG_Auth}}

\item{url}{API URL}

\item{timeo}{number of tries}
}
\value{
A \code{data.frame} of markets information
}
\description{
Fetch Markets Information
}
\examples{
HEADERS = IG_Auth(" ","APIdemo1", " ")
IG_fetch_markets(epic = "CS.D.AUDUSD.CFD.IP", HEADERS)

}
