% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAPI.R
\name{IG_fetch_open_pos}
\alias{IG_fetch_open_pos}
\title{IG API Returns all open positions for the active account.}
\usage{
IG_fetch_open_pos(headers,
  url = "https://demo-api.ig.com/gateway/deal/positions", timeo = 5)
}
\arguments{
\item{headers}{Object returned from \code{IG_Auth}}

\item{url}{API URL}

\item{timeo}{number of tries}
}
\value{
A \code{data.frame} open positions
}
\description{
Returns all open positions for the active account.
}
\examples{
HEADERS = IG_Auth(" ","APIdemo1", " ")
IG_fetch_open_pos(HEADERS)

}
