% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAPI.R
\name{IG_fetch_sub_nodes}
\alias{IG_fetch_sub_nodes}
\title{IG API Get Sub Navigation}
\usage{
IG_fetch_sub_nodes(headers,
  url = "https://demo-api.ig.com/gateway/deal/marketnavigation",
  node = 195235, timeo = 5)
}
\arguments{
\item{headers}{Object returned from \code{IG_Auth}}

\item{url}{API URL}

\item{node}{node id get from navigation}

\item{timeo}{number of tries}
}
\value{
A \code{data.frame} of node Navigation
}
\description{
Get Sub Navigation
}
\examples{
HEADERS = IG_Auth(" ","APIdemo1", " ")
IG_fetch_sub_nodes(HEADERS, node = 195235)

}
