% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAPI.R
\name{IG_fetch_top_navigation}
\alias{IG_fetch_top_navigation}
\title{IG API Get Top Navigation}
\usage{
IG_fetch_top_navigation(headers,
  url = "https://demo-api.ig.com/gateway/deal/marketnavigation", timeo = 5)
}
\arguments{
\item{headers}{Object returned from \code{IG_Auth}}

\item{url}{API URL}

\item{timeo}{number of tries}
}
\value{
A \code{data.frame} of Top Navigation
}
\description{
Get Top Navigation
}
\examples{
HEADERS = IG_Auth(" ","APIdemo1", " ")
IG_fetch_top_navigation(HEADERS)

}
