% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAPI.R
\name{IG_fetch_txns_hist}
\alias{IG_fetch_txns_hist}
\title{IG API Fetch Account Historical Transactions}
\usage{
IG_fetch_txns_hist(headers,
  url = "https://demo-api.ig.com/gateway/deal/history/transactions",
  trans_type = "ALL", milliseconds = 3600000, timeo = 5)
}
\arguments{
\item{headers}{Object returned from \code{IG_Auth}}

\item{url}{API URL}

\item{trans_type}{Types of transactions to retreive}

\item{milliseconds}{lookback periods}

\item{timeo}{number of tries}
}
\value{
A \code{data.frame} of account historical txns
}
\description{
Fetch IG Account Historical Transactions
}
\examples{
HEADERS = IG_Auth(" ","APIdemo1", " ")
IG_fetch_txns_hist(HEADERS)

}
