% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAPI.R
\name{IG_fetch_working_orders}
\alias{IG_fetch_working_orders}
\title{IG API Returns all open working orders for the active account.}
\usage{
IG_fetch_working_orders(headers,
  url = "https://demo-api.ig.com/gateway/deal/workingorders", timeo = 5)
}
\arguments{
\item{headers}{Object returned from \code{IG_Auth}}

\item{url}{API URL}

\item{timeo}{number of tries}
}
\value{
A \code{data.frame} List working orders response
}
\description{
Returns all open working orders for the active account.
}
\examples{
HEADERS = IG_Auth(" ","APIdemo1", " ")
IG_fetch_working_orders(HEADERS) # to be updated

}
