% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAPI.R
\name{IG_search_markets}
\alias{IG_search_markets}
\title{IG API Search Markets}
\usage{
IG_search_markets(headers,
  url = "https://demo-api.ig.com/gateway/deal/marketnavigation",
  search_term = "AUD", timeo = 5)
}
\arguments{
\item{headers}{Object returned from \code{IG_Auth}}

\item{url}{API URL}

\item{search_term}{search term}

\item{timeo}{number of tries}
}
\value{
A \code{data.frame} of markets information
}
\description{
Search Markets Information
}
\examples{
HEADERS = IG_Auth(" ","APIdemo1", " ")
IG_search_markets(HEADERS, search_term = "USD")

}
