% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAPI.R
\name{IG_update_open_pos}
\alias{IG_update_open_pos}
\title{IG API Updates an OTC position.}
\usage{
IG_update_open_pos(headers,
  url = "https://demo-api.ig.com/gateway/deal/positions/otc", limit_level,
  stop_level, deal_id, trailingStop, trailingStopIncrement,
  trailingStopDistance, timeo = 5)
}
\arguments{
\item{headers}{Object returned from \code{IG_Auth}}

\item{url}{API URL}

\item{limit_level}{Limit level}

\item{stop_level}{Stop level}

\item{deal_id}{deal identifier}

\item{trailingStop}{Whether the stop has to be moved towards the current level in case of a favourable trade}

\item{trailingStopIncrement}{increment step in pips for the trailing stop}

\item{trailingStopDistance}{Trailing stop distance}

\item{timeo}{number of tries}
}
\value{
A \code{data.frame} 	Deal reference of the transaction
}
\description{
Updates an OTC position.
}
\examples{
HEADERS = IG_Auth(" ","APIdemo1", " ")
IG_update_open_pos(HEADERS) # to be updated

}
