% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAPI.R
\name{IG_update_working_order}
\alias{IG_update_working_order}
\title{IG API Updates an OTC working order.}
\usage{
IG_update_working_order(headers,
  url = "https://demo-api.ig.com/gateway/deal/workingorders/otc",
  good_till_date, level, limit_distance, limit_level, stop_distance, stop_level,
  time_in_force, order_type, deal_id, timeo = 5)
}
\arguments{
\item{headers}{Object returned from \code{IG_Auth}}

\item{url}{API URL}

\item{good_till_date}{This accepts two possible formats either yyyy/mm/dd hh:mm:ss in UTC Time or Unix Timestamp in milliseconds}

\item{level}{Closing deal level}

\item{limit_distance}{Limit distance}

\item{limit_level}{Limit level}

\item{stop_distance}{Stop distance}

\item{stop_level}{Stop level}

\item{order_type}{'LIMIT', 'MARKET', 'QUATE'}

\item{deal_id}{deal ID}

\item{timeo}{number of tries}

\item{timeInForce}{'EXECUTE_AND_ELIMINATE' or 'FILL_OR_KILL'}
}
\value{
A \code{data.frame} 	OTC working order.
}
\description{
Updates an OTC working order.
}
\examples{
HEADERS = IG_Auth(" ","APIdemo1", " ")
IG_update_working_order(HEADERS) # to be updated

}
