% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQuantSQLServer.R
\name{InsitesSqlConnectJDBC}
\alias{InsitesSqlConnectJDBC}
\title{SQL Server Connection}
\usage{
InsitesSqlConnectJDBC(disconnect = FALSE, conn = NULL, firstTime = FALSE)
}
\arguments{
\item{disconnect}{TRUE/FALSE disconnect or connect}

\item{conn}{only requeired if you want to disconnect an existing connection}

\item{firstTime}{TRUE/FALSE if it's the first time to connect}
}
\value{
A \code{conn} of database
}
\description{
Connect to Sql Server databases
}
\examples{
# connect
conn <- InsitesSqlConnectJDBC()
sqlquery1 <- "select capiq_symbol_ticker, capiq_symbol_exchange, pricing_date, price_close, volume from insites_metrics_dev.meta.equity_price where pricing_date between GETDATE() - 500 and GETDATE()"
sqlquery2 <- "select distinct capiq_symbol_ticker, capiq_symbol_exchange from insites_metrics_dev.meta.equity_price where pricing_date between GETDATE() - 500 and GETDATE()"
res <- dbGetQuery(conn, sqlquery1)
res <- dbGetQuery(conn, sqlquery2)
sqlText <- paste("SELECT top 10 * FROM TABLE_NAME", sep="")
queryResults <- dbGetQuery(conn, sqlText)

# disconnect
InsitesSqlConnectJDBC(disconnect = TRUE, conn)

}
