% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuandlAPI.R
\name{QuandlConnect}
\alias{QuandlConnect}
\title{Quandl API Connection}
\usage{
QuandlConnect()
}
\description{
Register Quandl API key. For data retreival, please see \code{Quandl} packages for details
}
\examples{
QuandlConnect()
mydata = Quandl.datatable("ZACKS/FC", ticker="AAPL")

}
\seealso{
\link{Quandl.datatable}

\link{Quandl}
}
