% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBrokersInstrument.R
\name{reqIBFundamentalData}
\alias{reqIBFundamentalData}
\title{A Wrapper for IB Fundamental Data}
\usage{
reqIBFundamentalData(twsconn, reqId, contract, reportType)
}
\arguments{
\item{twsconn}{connection to current TWS account, if NULL, a new connection will be created within the function}

\item{reqId}{tws request ID}

\item{contract}{a twsContract}

\item{reportType}{can be: ReportsFinSummary, ReportsOwnership, ReportSnapshot, ReportsFinStatements, RESC (Analyst Estimates), CalendarReport}
}
\description{
Makes a request to the Interactive Brokers Trader Workstation (TWS), and returns the fundamental data.
}
\examples{
library(IBrokers)
tws <- twsConnect(port = 7497)
reqIBFundamentalData(tws, "1", twsSTK('AAPL'), 'ReportsFinSummary')

}
