% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBrokersInstrument.R
\name{reqIBNewsBulletins}
\alias{reqIBNewsBulletins}
\title{A Wrapper for IB News Bulletins Data}
\usage{
reqIBNewsBulletins(twsconn, allMsgs = TRUE)
}
\arguments{
\item{twsconn}{connection to current TWS account, if NULL, a new connection will be created within the function}

\item{allMsgs}{TRUE/FALSE}
}
\description{
Makes a request to the Interactive Brokers Trader Workstation (TWS), and returns the Bulletins data.
}
\examples{
library(IBrokers)
tws <- twsConnect(port = 7497)
reqNewsBulletins(tws, TRUE)

}
