% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwitterAPI.R
\name{setupTwitterConn}
\alias{setupTwitterConn}
\title{Sets up the OAuth credentials for a twitteR session}
\usage{
setupTwitterConn()
}
\description{
This function wraps the OAuth authentication handshake functions from the httr package for a twitteR session.
}
\examples{
setupTwitterConn()

}
