% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OandaForexLabs.R
\name{SpreadsOanda}
\alias{SpreadsOanda}
\title{Spread information for a supported instrument}
\usage{
SpreadsOanda(ACCOUNT_TYPE, ACCESS_TOKEN, ACCOUNT_ID, INSTRUMENTS = "EUR_USD",
  PERIOD = "1 year")
}
\arguments{
\item{ACCOUNT_TYPE}{"practice", "real" or "sandbox"}

\item{ACCESS_TOKEN}{The authorization bearer token previously obtained by the client, can be found in oanda environment}

\item{ACCOUNT_ID}{ID of the Account to create the Order for.}

\item{INSTRUMENTS}{Instrument to get candlestick data for e.g. "AUD_USD"}

\item{PERIOD}{Required Period of time to retrieve calendar data for.}
}
\value{
A \code{list}
}
\description{
Returns up to 1 year worth of spread information for a supported instrument.
The returned data is divided in 15 minute intervals.
For each period, we provide the time weighted average, minimum, and maximum spread.
}
\details{
Period includes:\cr
 - 1 hour\cr
 - 12 hours\cr
 - 1 day\cr
 - 1 week\cr
 - 1 month\cr
 - 3 months\cr
 - 6 months\cr
 - 1 year\cr
}
\examples{
spreads = SpreadsOanda(.oandaEnv$ACCOUNT_TYPE, .oandaEnv$ACCESS_TOKEN, .oandaEnv$ACCOUNT_ID, INSTRUMENTS = 'AUD_USD', PERIOD = '1 year')
spreads

}
