% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQuantSQLServer.R
\name{sqlServerConnectODBC}
\alias{sqlServerConnectODBC}
\title{SQL Server Connection}
\usage{
sqlServerConnectODBC()
}
\value{
A \code{conn} of database
}
\description{
Connect to Sql Server databases
}
\examples{
conn <- sqlServerConnect()
sqlquery1 <- "select capiq_symbol_ticker, capiq_symbol_exchange, pricing_date, price_close, volume from insites_metrics_dev.meta.equity_price where pricing_date between GETDATE() - 500 and GETDATE()"
sqlquery2 <- "select distinct capiq_symbol_ticker, capiq_symbol_exchange from insites_metrics_dev.meta.equity_price where pricing_date between GETDATE() - 500 and GETDATE()"
res <- sqlFetch(conn, sqlquery1)

}
