% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwitterAPI.R
\name{tweet_corpus}
\alias{tweet_corpus}
\title{Submit a search query (terms separated by "+") and get a return}
\usage{
tweet_corpus(search, n = 5000, since = NULL, until = NULL,
  sinceID = NULL, maxID = NULL, ...)
}
\arguments{
\item{search}{Search query to issue to twitter. Use "+" to separate query terms.}

\item{n}{The maximum number of tweets to return}

\item{since}{If not NULL, restricts tweets to those since the given date. Date is to be formatted as YYYY-MM-DD}

\item{until}{If not NULL, restricts tweets to those up until the given date. Date is to be formatted as YYYY-MM-DD}

\item{sinceID}{If not NULL, returns tweets with IDs greater (ie newer) than the specified ID}

\item{maxID}{If not NULL, returns tweets with IDs smaller (ie older) than the specified ID}
}
\value{
A \code{list} containing a vector of all doc and a data.frame of all tweets
}
\description{
Submit a search query (terms separated by "+") and get a return set of data (corpus).
}
\examples{
setupTwitterConn()
tweets <- tweet_corpus(search = "audusd OR aud_usd OR aud/usd", n = 1000, since = as.character(Sys.Date()-1), until = as.character(Sys.Date()))

}
