% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwitterAPI.R
\name{tweet_corpus_chunck}
\alias{tweet_corpus_chunck}
\title{Submit a search query by chuncks}
\usage{
tweet_corpus_chunck(search = "aud+usd", n = 50000, chunck_size = 7,
  total_size = 15, until = Sys.Date(), sys_sleep = 0)
}
\arguments{
\item{search}{Search query to issue to twitter. Use "+" to separate query terms.}

\item{n}{The maximum number of tweets to return}

\item{chunck_size}{number of days for each iteration}

\item{total_size}{total number of days to retrieve}

\item{until}{If not NULL, restricts tweets to those up until the given date. Date is to be formatted as YYYY-MM-DD}

\item{sys_sleep}{system sleep (sec) to prevent limits rate of API}
}
\value{
A \code{list} containing a vector of all doc and a data.frame of all tweets
}
\description{
Submit a search query by chuncks
}
\examples{
tweets <- tweet_corpus_chunck(search = "aud+usd", n = 50000, chunck_size = 7, total_size = 15, until = Sys.Date())

}
