% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwitterAPI.R
\name{tweet_text}
\alias{tweet_text}
\title{Convenience function for accessing the text part of a tweet}
\usage{
tweet_text(x)
}
\description{
Convenience function for accessing the text part of a tweet returned by the twitteR API.
}
