% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwitterAPI.R
\name{twitterExtracter}
\alias{twitterExtracter}
\title{Extract FX Tweets}
\usage{
twitterExtracter(inst1, inst2)
}
\arguments{
\item{inst1}{instrument one}

\item{inst2}{instrument two}
}
\description{
Extract Twitter by days and save the data in specified folder
}
\examples{
twitterExtracter("AUD", "USD")

}
